Create stock_bot
A more engaged bot to allow a user to input a stock name and it will pull up relevant ticker information.  It will include several delta computations to determine the percent change of the price per day/week/two week over a date the user inputs.

As a project this is designed to increase my ability to work with and understand OOP concepts, lists, and exporting to a spreadsheet, some SQL, and ideally I'd like to use flask to export it to an HTML file and create a Google Graphs file and a functioning application.